refactor: consolidate getMessageDir (null-returning) to shared/session-utils
