+++ src/features/background-agent/message-dir.ts
@@ -1,1 +1,1 @@
-export { getMessageDir } from "./message-storage-locator"
+export { getMessageDir } from "../../shared/session-utils"
