--- src/tools/call-omo-agent/message-storage-directory.ts
@@ -1,18 +1,0 @@
-import { existsSync, readdirSync } from "node:fs"
-import { join } from "node:path"
-import { MESSAGE_STORAGE } from "../../features/hook-message-injector"
-
-export function getMessageDir(sessionID: string): string | null {
-	if (!sessionID.startsWith("ses_")) return null
-	if (!existsSync(MESSAGE_STORAGE)) return null
-
-	const directPath = join(MESSAGE_STORAGE, sessionID)
-	if (existsSync(directPath)) return directPath
-
-	for (const dir of readdirSync(MESSAGE_STORAGE)) {
-		const sessionPath = join(MESSAGE_STORAGE, dir, sessionID)
-		if (existsSync(sessionPath)) return sessionPath
-	}
-
-	return null
-}
